rename spawner classes
